tools/molecular_dynamics_xtb/requirements.txt
# Tool dependencies
fastmcp
rdkit>=2023.9.1
numpy>=1.24.0
scipy>=1.11.0
pillow>=9.0.0
imageio>=2.0.0
